feat(viriditas): adopt Sprinkles, and begin building a polymorphic Box component
so far, Sprinkles is only used for text stuff, since those are the only properties
I've adopted systematic, constraint-based design for at this time.
for the Box component, I referenced these two approaches:
 - https://sandroroth.com/blog/vanilla-extract-approach/
 - https://www.highlight.io/blog/typesafe-tailwind
mine is a bit less full-featured than either of those. but it works, and that was
the main hurdle for today!